% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoExploreRPlots.R
\name{allDistancesWalk}
\alias{allDistancesWalk}
\title{plot distances}
\usage{
allDistancesWalk(ds, ymin = 0, ymax = 35)
}
\arguments{
\item{ds}{is a list of distances}
}
\description{
plot distances
}
