% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoExploreRPlots.R
\name{allGenesWalk}
\alias{allGenesWalk}
\title{plots walks of select genes in the same plot}
\usage{
allGenesWalk(x, indrange = 1:10, ymin = -40, ymax = 40)
}
\arguments{
\item{x}{is a matrix of gene walks with genes on rows and time steps on columns}

\item{indrange}{specifies which genes to plot}

\item{ncol}{specifies how many columns to wrap on}
}
\description{
plots walks of select genes in the same plot
}
