% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoExploreRPlots.R
\name{allPathwaysWalk}
\alias{allPathwaysWalk}
\title{plots walks of select pathways in the same plot}
\usage{
allPathwaysWalk(y, indrange = 1:10, ymin = -1, ymax = 1)
}
\arguments{
\item{y}{is a matrix of pathway walks with pathways on rows and time steps on columns}

\item{indrange}{is the indices of pathways to plot}

\item{ncol}{specifies how many columns to wrap on}
}
\description{
plots walks of select pathways in the same plot
}
