% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoExploreRWalkers.R
\name{checkDataType}
\alias{checkDataType}
\title{Checks that data is in the right format}
\usage{
checkDataType(expr)
}
\arguments{
\item{expr}{Gene Expression dataset}
}
\value{
expr Gene Expression dataset in the proper form (may be the same as input)
}
\description{
Checks that data is in the right format
}
\examples{
expr.matrix <- matrix(rnorm(100),ncol=15)
checkDataType(expr.matrix)
#TRUE

}
