% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoExploreRDistances.R
\name{computePDD}
\alias{computePDD}
\title{Pathway Distribution Distance}
\usage{
computePDD(m, design, min = -1, max = 1, breaks = 50,
  asGDD = FALSE)
}
\arguments{
\item{m}{a matrix of pathway activities, rows are pathways, columns samples}

\item{design}{a vector of labels with phenotypes}

\item{min}{minimum value for histogram scale; default -1, based on activity values}

\item{max}{maximum value for histogram scale; default 1, based on activity values}

\item{breaks}{a number of breaks to use to create pathway distribution histograms}

\item{asGDD}{whether to return the summary as Global Distribution Distance; summed over PDDs}
}
\value{
Pathway Distribution Distance
}
\description{
Pathway Distribution Distance
}
