% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoExploreRPlots.R
\name{distributionsWalk}
\alias{distributionsWalk}
\title{plots distribution walks of select pathways}
\usage{
distributionsWalk(w, startLoc, destLoc, pheno, indrange, samplefreq = 50,
  ncol = 5)
}
\arguments{
\item{w}{is a list of pathway walks matrices}

\item{startLoc}{specifies which is the start location}

\item{destLoc}{specifies target location}

\item{pheno}{specifies the phenotype object}

\item{indrange}{specifies the indices of pathways to plot}

\item{samplefreq}{specifies at what frequency to sample the walks}

\item{ncol}{specifies how many columns to wrap on}
}
\description{
plots distribution walks of select pathways
}
