% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoExploreRWalkers.R
\name{explore}
\alias{explore}
\title{Phenotype Explorer}
\usage{
explore(expr, design, motif, pathways, logTransform = TRUE,
  maxIter = 100, percent = 0.8, Dsquare = 0.1, deltaT = 4,
  alpha = 0.5, Jscaling = 1, outputDir = NA, numMaxCores = 1)
}
\arguments{
\item{expr}{Gene Expression dataset, can be matrix or data.frame of expression values}

\item{design}{Binary vector indicating init and dest location information}

\item{motif}{Regulatory data, can be a binary square matrix or data frame where motif[i,j]=1 indicates TF i regulates gene j}

\item{pathways}{list of genesets to use of phenotype}

\item{logTransform}{indicates whether log transformation of expression is done initally}

\item{maxIter}{numer of steps to simulate}

\item{percent}{the proportion of the destination distributon (centered around the mean) that should be reached for convergence}

\item{Dsquare}{noise scaling parameter}

\item{deltaT}{time scaling parameter}

\item{alpha}{step scaling parameter}

\item{Jscaling}{indicates whether a gradient descent is taken: Jscaling 0 < 1 achieves for gradient descent; Jscaling=1 for no scaling}

\item{outputDir}{character vector specifying a directory where to save walker results, default is NA}

\item{numMaxCores}{requires doParallel, foreach.  Runs walker in parallel computing environment.  Set to 1 to avoid parallelization.}
}
\value{
An object of class "phenoExplorer" containing results
}
\description{
Phenotype Explorer
}
\examples{
data(IVY)
design <- c(rep(0, 10), rep(NA,20),rep(1,10))
phenoExplorerRes <- explore(IVY$expr[,0:40],design,IVY$motif,IVY$pathways,maxIter=10, outputDir="simulations",numMaxCores=4)
}
