% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoExploreRPlots.R
\name{genesWalk}
\alias{genesWalk}
\title{plots walks of select individual genes}
\usage{
genesWalk(x, indrange = 1:10, ncol = 5, scales = "free")
}
\arguments{
\item{x}{is a matrix of gene walks, with genes on rows and time steps on columns}

\item{indrange}{specifies which genes to plot}

\item{ncol}{specifies how many columns to wrap on}

\item{scales}{specifies whether to have "fixed" or "free" scale}
}
\description{
plots walks of select individual genes
}
