% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoExploreRPlots.R
\name{pathwaysWalk}
\alias{pathwaysWalk}
\title{plots walks of select indivdual pathways}
\usage{
pathwaysWalk(y, indrange = 1:10, bounds, ncol = 5, scales = "free")
}
\arguments{
\item{y}{is a matrix of pathway walks with pathways on rows and time steps on columns}

\item{indrange}{is the indices of pathways to plot}

\item{bounds}{is the quantile of target destination}

\item{ncol}{specifies how many columns to wrap on}

\item{scales}{whether to have "fixed" or "free" scales}
}
\description{
plots walks of select indivdual pathways
}
