% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoExploreRDistances.R
\name{permuteDesign}
\alias{permuteDesign}
\title{Design permutation}
\usage{
permuteDesign(init.design, permutation.n = 1000,
  permutation.ratio = c(0.4, 0.6))
}
\arguments{
\item{init.design}{a vector of labels with two phenotypes}

\item{permutation.n}{a number of permutations}

\item{permuation.ratio}{interval specifiying the percentage of overlap with original design}
}
\value{
permutations of labels
}
\description{
Design permutation
}
