% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoExploreRPlots.R
\name{plotPhenotypePCA}
\alias{plotPhenotypePCA}
\title{p is a phenotype object}
\usage{
plotPhenotypePCA(pheno, expr = TRUE, method = c("PCA", "BGA"))
}
\arguments{
\item{expr}{is a flag for whether to plot gene level (TRUE) or pathway activity level}

\item{method}{is a choice of "PCA" or "BGA"; default is PCA}
}
\description{
p is a phenotype object
}
